+++ CHANGES.rst
@@ -144,6 +144,9 @@ tweakreg
   of the 'GAIA' catalogs for aligning all input images together as one single
   group. [#6946]
 
+- exposed additional parameters for absolute astrometry:
+ ``abs_minobj``, ``abs_searchrad``, ``abs_use2dhist``, ``abs_separation``, ``abs_tolerance``, ``abs_fitgeometry``, ``abs_nclip``,  and ``abs_sigma``. [#6987]
+
 - Refactored code to work with changes in ``tweakwcs`` version 0.8.0. [#7006]